# Extract and optimize group member calculation