--- a/pkg/controller/networkpolicy/store/group.go
+++ b/pkg/controller/networkpolicy/store/group.go
@@ -27,8 +27,7 @@ import (
 )
 
 const (
-	GroupMemberIndex = "groupMember"
-	ServiceIndex     = "service"
+	ServiceIndex = "service"
 )
 
 // GroupKeyFunc knows how to get the key of an Group.
@@ -50,21 +49,6 @@ func NewGroupStore() storage.Interface {
 				return []string{}, nil
 			}
 			return []string{g.Selector.Namespace}, nil
-		},
-		GroupMemberIndex: func(obj interface{}) ([]string, error) {
-			g, ok := obj.(*antreatypes.Group)
-			if !ok {
-				return []string{}, nil
-			}
-			var members []string
-			for _, m := range g.GroupMembers {
-				if m.Pod != nil {
-					members = append(members, k8s.NamespacedName(m.Pod.Namespace, m.Pod.Name))
-				} else if m.ExternalEntity != nil {
-					members = append(members, k8s.NamespacedName(m.ExternalEntity.Namespace, m.ExternalEntity.Name))
-				}
-			}
-			return members, nil
 		},
 		ServiceIndex: func(obj interface{}) ([]string, error) {
 			g, ok := obj.(*antreatypes.Group)