Previously grouping member calculation was coupled with the NetworkPolicyController, and there were two kinds of heavy calculation in the process:
For each event like Pod add, it needed to find all affected groups by matching the new Pod's labels with all groups that can potentially select it. If it was an update event, it even needed to perform the process one more time for its old labels. Then it triggered the sync process of the affected groups.
For the sync process of a group, it needed to find all Pods or ExternalEntities by matching its selectors with all entities that can potentially match it.
For each pair of Pod and Group, they would be matched at least twice. Besides, whenever the group was processed another time, it would again scan all Pods.

The repeated calculation can be eliminated by caching the result of the first matching process when processing Pod events. Then the matching process will only occur when there's a new Pod or a new Group.
Besides, most Pods actually share same labels with others when they are managed by same controller. The cache can be further optimized to maintain the relationship between label set and label selector, making Pods that having same labels and Groups that having same selectors share the cache.

This patch introduces an interface which is responsible for group member calculation and an implementation implementing it with the above optimization. The interface is then consumed by NetworkPolicyController and features like EndpointQuerier and ClusterGroup to retrieve Pods/ExternalEntities selected by a given Group or Groups that select a given Pod. Besides, other controllers that have the group logic like EgressPolicy can consume it directly without having redundant code with NetworkPolicyController.

Performance impact:
```
name                                       old time/op    new time/op    delta
InitXLargeScaleWithSmallNamespaces-48         3.65s ±14%     4.35s ± 7%  +19.20%  (p=0.016 n=5+5)
InitXLargeScaleWithLargeNamespaces-48         5.34s ± 2%     1.48s ± 5%  -72.34%  (p=0.008 n=5+5)
InitXLargeScaleWithOneNamespace-48            3.22s ± 8%     3.46s ± 2%     ~     (p=0.056 n=5+5)
InitXLargeScaleWithNetpolPerPod-48            45.9s ±11%     25.6s ± 4%  -44.18%  (p=0.008 n=5+5)
InitXLargeScaleWithClusterScopedNetpol-48     1.40s ± 3%     1.18s ± 5%  -16.08%  (p=0.008 n=5+5)

name                                       old alloc/op   new alloc/op   delta
InitXLargeScaleWithSmallNamespaces-48         824MB ± 0%     900MB ± 0%   +9.22%  (p=0.008 n=5+5)
InitXLargeScaleWithLargeNamespaces-48         629MB ± 0%     321MB ± 0%  -49.00%  (p=0.008 n=5+5)
InitXLargeScaleWithOneNamespace-48            921MB ± 0%     922MB ± 0%   +0.13%  (p=0.008 n=5+5)
InitXLargeScaleWithNetpolPerPod-48           3.39GB ± 0%    0.13GB ± 0%  -96.19%  (p=0.008 n=5+5)
InitXLargeScaleWithClusterScopedNetpol-48     279MB ± 0%     244MB ± 0%  -12.59%  (p=0.008 n=5+5)

name                                       old allocs/op  new allocs/op  delta
InitXLargeScaleWithSmallNamespaces-48         12.0M ± 0%     12.6M ± 0%   +5.35%  (p=0.008 n=5+5)
InitXLargeScaleWithLargeNamespaces-48         4.29M ± 0%     4.48M ± 0%   +4.63%  (p=0.008 n=5+5)
InitXLargeScaleWithOneNamespace-48            1.39M ± 0%     1.40M ± 0%   +0.74%  (p=0.008 n=5+5)
InitXLargeScaleWithNetpolPerPod-48            1.66M ± 0%     1.75M ± 0%   +5.52%  (p=0.008 n=5+5)
InitXLargeScaleWithClusterScopedNetpol-48     3.07M ± 0%     3.20M ± 0%   +4.04%  (p=0.008 n=5+5)
```

Explanation:
- `InitXLargeScaleWithSmallNamespaces` has only 4 Pods in each Namespace and all NetworkPolicies are namespace scoped, so even without the index, each policy just needs to scan 4 Pods at most. Maintaing the index adds few overhead, hence the minor increment in time and memory.
- `InitXLargeScaleWithLargeNamespaces` has 1000 Pods and 100 NetworkPolicies in each Namespace, each NetworkPolicy applies to 10 Pods. Pods that have same labels will only be scaned once so many calculation is saved. When syncing groups, the result can be got directly without listing all Pods in the Namespace first, so calculation and memory is saved.
- `InitXLargeScaleWithOneNamespace` just has many Pods but only 1 namespace, 1 AppliedToGroup and 1 AddressGroup in total, so having the index doesn't help anything.
- `InitXLargeScaleNetpolPerPod` has only 1 namespace, 10000 Pods and 1 NetworkPolicy per Pod. It can benefit from the index greatly for the same reason as `InitXLargeScaleWithLargeNamespaces`.