@tnqn 2021-03-03
@jianjuns @antoninbas @Dyanngg Please let me know if the change makes sense to you. And whether it can help nested group feature or make it harder? @Dyanngg 
I will add performance impact later.