@Dyanngg 2021-03-03
> @jianjuns @antoninbas @Dyanngg Please let me know if the change makes sense to you. And whether it can help nested group feature or make it harder? @Dyanngg
> I will add performance impact later.

I'm still reviewing the `GroupEntityIndex` interface, but after the first look I don't think it will make nested group computation easier or more complicated. It's just that the nested group member calc logic would be entirely re-written, since right now I rely on the `groupMember` field of the internalGroup to union nested groupMembers, which is removed by this PR.