@tnqn 2021-03-04
> > @jianjuns @antoninbas @Dyanngg Please let me know if the change makes sense to you. And whether it can help nested group feature or make it harder? @Dyanngg
> > I will add performance impact later.
> 
> I'm still reviewing the `GroupEntityIndex` interface, but after the first look I don't think it will make nested group computation easier or more complicated. It's just that the nested group member calc logic would be entirely re-written, since right now I rely on the `groupMember` field of the internalGroup to union nested groupMembers, which is removed by this PR.

I removed the field as it was consumed by `GetGroupMembers`, `populateAddressGroupMemberSet` and `getAppliedToWorkloads` when they need to know the members of a clustergroup but with the change it can be achieved by querying the `GroupEntityIndex` directly, so the internalGroup itself doesn't have to store the redundant data. I can imagine that in nested group case, you want to avoid calculate the union of the members again when querying its member, right? If I understand correctly, I can add this field back or let's see whether there is more efficient way.