@tnqn 2021-03-16
@jianjuns @antoninbas @Dyanngg I added more unit tests and a method `HasSynced` to avoid intermediate result after restarting since your last review and have addressed all comments if I'm not missing one. Could you take another look?